\name{GOFunction-package}
\alias{GOFunction-package}

\docType{package}
\title{GO-function: deriving biologically relevant functions from statistically significant functions}

\description{
    The GO-function package is an enrichment analysis tool for Gene Ontology (GO). According to some explicit rules, 
    it is designed for treating the redundancy resulting from the GO structure or multiple annotation genes. Different 
    from current redundancy treatment tools simply based on some numerical considerations, GO-function can find terms 
    which are both statistically interpretable and biologically meaningful.
}
\details{
  
  \tabular{ll}{
      Package: \tab GOFunction\cr
      Type: \tab Package\cr
      Version: \tab 1.0.1\cr
      Date: \tab 2011-06-14\cr
      License: \tab GPL (>=2)\cr
      LazyLoad: \tab yes\cr
   }
}

\author{
   Jing Wang
   Maintainer: Jing Wang <jingwang.uestc@gmail.com>
}

\references{
    Jing Wang et al. GO-function: deriving biologically relevant functions from statistically significant functions, Brief Bioinform, 2012, 13(2): 216-227
}

\seealso{
    \code{\link{GOFunction}}
}

\keyword{package}