\name{interestGenes}
\alias{interestGenes}
\docType{data}
\title{
   Colon cancer differentially expressed genes
}
\description{
   The interestGenes contain 9201 differentially expressed genes extracted from Sabates-Bellver dataset.
}
\usage{data(exampledata)}
\format{vector}
\details{
   The differentially expressed genes are selected using the Significant Analysis of Microarray (SAM) with an FDR of 1%.
}
\references{
   Sabates-Bellver J. et al. Transcriptome prole of human colorectal adenomas. Mol Cancer Res, 5(12):1263-1275, 2007.
}
\keyword{datasets}
