\name{refGenes}
\alias{refGenes}
\docType{data}
\title{
   Colon cancer background genes
}
\description{
   The refGenes contain 19950 background genes.
}
\usage{data(exampledata)}
\format{vector}
\references{
   Sabates-Bellver J. et al. Transcriptome prole of human colorectal adenomas. Mol Cancer Res, 5(12):1263-1275, 2007.
}
\keyword{datasets}
